Remove race condition
When an ext irq interrupts HS it performs plic_claim and stores
the id into a variable. This variable is only used by the current
cpu until the check for (phys_irq == this_variable).
Then it is first zeroed and only after that plic_complete is called.
It is like PLIC's mechanisms are used for locking this variable.